Tweaked escorts and added stealth mechanics.
The escorts for the Heisenberg and the Thassos will warp out when the
ships get close to the Regulus node, which gives the player a short
window to dock with the Heisenberg safely. I also added logic for making
the NTF detect the player by visual contact, or when he fires weapons or
uses afterburners too close to enemies.