Implemented sensor decoy mechanic.
Instead of having the Heisenberg's escorts conveniently jump out to give
the player a chance to dock undetected, I added a way for the player to
have more agency. The player now can partially cut power to the Pegasus'
stealth system for an instant to create a faint sensor blip that will
lure NTF escorts away. When the player presses a button, the nearest NTF
escort wing will break formation and go to wherever the player was when
the button was pressed.

This is a basic implementation of this mechanic, and it needs to be
refined later. Particularly: making escorts turn back if they get some
distance away from their original position, even if they haven't reached
the waypoint.