Improved sneaky docking, and a couple table fixes.
I added new dock points on the back end of the Faustus model and on the
underside of the Pegasus model. This way, when you sneakily dock with
the Heisenberg, it looks like you just clamp onto a random part of the
ship's hull instead of parking on one of its airlocks.