Finished initial draft of mission 7 (parts A and B)
The SD Yama and the SCv Narada pursue the Phantom to an unknown gas
giant planet. To escape, the Phantom is forced to jump into the gas
giant's upper atmosphere. This seems to make the Shivans give up their
pursuit, but it soon becomes clear that the Phantom is being shadowed by
stealthy Shivan fighters, which for reasons unknown, are not openly
hostile...